param `auto_init` for repo and fix race conditions for branch protection (#111)
* param `auto_init` for repo and fix race conditions for branch protection

Basically when you create a new repo you can have a race condition where it wants to create the branch protection. The problem is that you have two dependencies:
- the repo itself
- the branch to be protected needs to exist

By adding an explicit depends we can ensure that only after the repo resource has been run will it attempt to setup branch protections.

Also in the case of wanting to create a repository and copy the history of another project in you can specify `auto_init` and it will not attempt to initialize the repo with a commit or setup protection. After copying the history you can flip `auto_init` to `true` which will flip the branch protection on without creating any commits.

* work around GH provider bug with `auto_init` value being changed destroys a repo.
* if we set a repo to be archived we should change the permissions from whatever we specified to pull and otherwise fall back to what we specified. This could be accomplished as well by setting the various team permissions but that seems like more effort and error prone than simply baking in the expected behavior and expect people to only set 1 thing instead of 3.

Signed-off-by: Ben Abrams <me@benabrams.it>